feat: 소켓 인스턴스 설정하는 위치 수정
